Extract _fonts.py + _overlays.py from _legacy.py
_legacy.py: -458 lines (2837 remaining).
