+++ modules/helpers/__init__.py
@@ -27,4 +27,6 @@ from ._legacy import (  # noqa: F401
 # will override legacy definitions (in case of a future rename).
 from ._cli import *  # noqa: F403
 from ._templates import *  # noqa: F403
+from ._fonts import *  # noqa: F403
+from ._overlays import *  # noqa: F403
 from ._paths import *  # noqa: F403
