+++ modules/helpers/_legacy.py
@@ -424,30 +424,6 @@ def extract_library_name(key):
     return match.group(1) if match else None
 
 
-def get_pyfiglet_fonts():
-    """Retrieve available PyFiglet fonts from static/fonts, sorted with custom order."""
-    if getattr(sys, "frozen", False):  # running in frozen/packaged mode
-        base_path = sys._MEIPASS
-    else:
-        base_path = os.path.abspath(".")
-
-    fonts_dir = os.path.join(base_path, "static", "fonts")
-
-    # Ensure predefined fonts are at the top
-    predefined_fonts = ["none", "single line", "standard"]
-    fonts = set(predefined_fonts)  # Using set to prevent duplicates
-
-    # Append all .flf files, removing extension
-    if os.path.exists(fonts_dir):
-        fonts.update(f.replace(".flf", "") for f in os.listdir(fonts_dir) if f.endswith(".flf"))
-
-    # Sort remaining fonts (excluding predefined ones)
-    sorted_fonts = sorted(fonts - set(predefined_fonts))
-
-    # Combine predefined fonts with sorted remaining fonts
-    return predefined_fonts + sorted_fonts
-
-
 def calculate_hash(content):
     """Compute the SHA256 hash of the given content."""
     return hashlib.sha256(content.encode("utf-8")).hexdigest()
@@ -905,262 +881,6 @@ def consume_restart_notice():
     except Exception as exc:
         ts_log(f"Failed to remove restart notice: {exc}", level="WARNING")
     return payload
-
-
-def load_quickstart_config(filename: str):
-    json_path = os.path.join(JSON_SETTINGS, filename)
-    with open(json_path, "r", encoding="utf-8") as f:
-        return json.load(f)
-
-
-def _overlay_origin_alignment_defaults(origin):
-    origin_str = str(origin or "").strip().lower()
-    tokens = [token for token in re.split(r"[^a-z]+", origin_str) if token]
-    has_center = "center" in tokens
-
-    horizontal = "right" if "right" in tokens else "left"
-    if "left" not in tokens and has_center and "right" not in tokens:
-        horizontal = "center"
-
-    vertical = "bottom" if "bottom" in tokens else "top"
-    if "top" not in tokens and has_center and "bottom" not in tokens:
-        vertical = "center"
-
-    return horizontal, vertical
-
-
-def _overlay_template_var_keys(template_variables):
-    keys = set()
-    if isinstance(template_variables, dict):
-        keys.update(str(key) for key in template_variables.keys())
-    elif isinstance(template_variables, list):
-        for item in template_variables:
-            if isinstance(item, dict) and item.get("key"):
-                keys.add(str(item.get("key")))
-    return keys
-
-
-def _insert_overlay_mapping_fields(template_variables, inserted_fields, *, before_keys=None):
-    if not isinstance(template_variables, dict) or not inserted_fields:
-        return template_variables
-
-    before_keys = set(before_keys or ())
-    next_template_variables = {}
-    inserted = False
-
-    for key, value in template_variables.items():
-        if not inserted and key in before_keys:
-            for inserted_key, inserted_value in inserted_fields.items():
-                if inserted_key not in template_variables:
-                    next_template_variables[inserted_key] = inserted_value
-            inserted = True
-        next_template_variables[key] = value
-
-    if not inserted:
-        for inserted_key, inserted_value in inserted_fields.items():
-            if inserted_key not in next_template_variables:
-                next_template_variables[inserted_key] = inserted_value
-
-    template_variables.clear()
-    template_variables.update(next_template_variables)
-    return template_variables
-
-
-def _insert_overlay_list_fields(template_variables, inserted_fields, *, before_keys=None):
-    if not isinstance(template_variables, list) or not inserted_fields:
-        return template_variables
-
-    before_keys = set(before_keys or ())
-    existing_keys = _overlay_template_var_keys(template_variables)
-    next_template_variables = []
-    inserted = False
-
-    for item in template_variables:
-        item_key = item.get("key") if isinstance(item, dict) else None
-        if not inserted and item_key in before_keys:
-            for inserted_field in inserted_fields:
-                inserted_key = inserted_field.get("key")
-                if inserted_key and inserted_key not in existing_keys:
-                    next_template_variables.append(copy.deepcopy(inserted_field))
-            inserted = True
-        next_template_variables.append(item)
-
-    if not inserted:
-        for inserted_field in inserted_fields:
-            inserted_key = inserted_field.get("key")
-            if inserted_key and inserted_key not in existing_keys:
-                next_template_variables.append(copy.deepcopy(inserted_field))
-
-    template_variables[:] = next_template_variables
-    return template_variables
-
-
-def enrich_quickstart_overlay_config(config):
-    enriched = copy.deepcopy(config or [])
-
-    for group in enriched:
-        if not isinstance(group, dict):
-            continue
-        overlays = group.get("overlays", [])
-        if not isinstance(overlays, list):
-            continue
-        for overlay in overlays:
-            if not isinstance(overlay, dict):
-                continue
-            template_variables = overlay.get("template_variables")
-            if template_variables is None:
-                template_variables = {}
-                overlay["template_variables"] = template_variables
-            if not isinstance(template_variables, (dict, list)):
-                continue
-
-            existing_keys = _overlay_template_var_keys(template_variables)
-            default_offsets = overlay.get("default_offsets")
-            offsets_by_type = overlay.get("default_offsets_by_type")
-
-            supports_runtime_offsets = (
-                isinstance(default_offsets, dict)
-                or (isinstance(offsets_by_type, dict) and any(isinstance(value, dict) for value in offsets_by_type.values()))
-                or "initial_horizontal_offset" in existing_keys
-                or "initial_vertical_offset" in existing_keys
-            )
-
-            offset_defaults = default_offsets if isinstance(default_offsets, dict) else {}
-            if not offset_defaults and isinstance(offsets_by_type, dict):
-                for candidate in offsets_by_type.values():
-                    if isinstance(candidate, dict):
-                        offset_defaults = candidate
-                        break
-
-            if supports_runtime_offsets:
-                horizontal_default = 0
-                vertical_default = 0
-                if "initial_horizontal_offset" in existing_keys:
-                    if isinstance(template_variables, dict):
-                        horizontal_default = template_variables.get("initial_horizontal_offset", {}).get(
-                            "default",
-                            offset_defaults.get("horizontal", 0),
-                        )
-                    else:
-                        for item in template_variables:
-                            if isinstance(item, dict) and item.get("key") == "initial_horizontal_offset":
-                                horizontal_default = item.get("default", offset_defaults.get("horizontal", 0))
-                                break
-                else:
-                    horizontal_default = offset_defaults.get("horizontal", 0)
-
-                if "initial_vertical_offset" in existing_keys:
-                    if isinstance(template_variables, dict):
-                        vertical_default = template_variables.get("initial_vertical_offset", {}).get(
-                            "default",
-                            offset_defaults.get("vertical", 0),
-                        )
-                    else:
-                        for item in template_variables:
-                            if isinstance(item, dict) and item.get("key") == "initial_vertical_offset":
-                                vertical_default = item.get("default", offset_defaults.get("vertical", 0))
-                                break
-                else:
-                    vertical_default = offset_defaults.get("vertical", 0)
-
-                offset_fields_mapping = {
-                    "horizontal_offset": {
-                        "input_type": "number",
-                        "default": horizontal_default,
-                        "label": "Horizontal Offset",
-                    },
-                    "vertical_offset": {
-                        "input_type": "number",
-                        "default": vertical_default,
-                        "label": "Vertical Offset",
-                    },
-                }
-                offset_fields_list = [
-                    {
-                        "input_type": "number",
-                        "key": "horizontal_offset",
-                        "default": horizontal_default,
-                        "label": "Horizontal Offset",
-                    },
-                    {
-                        "input_type": "number",
-                        "key": "vertical_offset",
-                        "default": vertical_default,
-                        "label": "Vertical Offset",
-                    },
-                ]
-
-                if isinstance(template_variables, dict):
-                    _insert_overlay_mapping_fields(
-                        template_variables,
-                        offset_fields_mapping,
-                        before_keys={"horizontal_offset", "vertical_offset", "builder_level"},
-                    )
-                else:
-                    _insert_overlay_list_fields(
-                        template_variables,
-                        offset_fields_list,
-                        before_keys={"horizontal_offset", "vertical_offset", "builder_level"},
-                    )
-
-            supports_origin_alignment = (
-                isinstance(offset_defaults, dict)
-                and bool(offset_defaults.get("origin"))
-                and "horizontal_position" not in existing_keys
-                and "vertical_position" not in existing_keys
-            )
-
-            if supports_origin_alignment:
-                horizontal_align_default, vertical_align_default = _overlay_origin_alignment_defaults(offset_defaults.get("origin"))
-                align_fields_mapping = {
-                    "horizontal_align": {
-                        "input_type": "select",
-                        "default": horizontal_align_default,
-                        "label": "Horizontal Alignment",
-                        "options": ["left", "center", "right"],
-                    },
-                    "vertical_align": {
-                        "input_type": "select",
-                        "default": vertical_align_default,
-                        "label": "Vertical Alignment",
-                        "options": ["top", "center", "bottom"],
-                    },
-                }
-                align_fields_list = [
-                    {
-                        "input_type": "select",
-                        "key": "horizontal_align",
-                        "default": horizontal_align_default,
-                        "label": "Horizontal Alignment",
-                        "options": ["left", "center", "right"],
-                    },
-                    {
-                        "input_type": "select",
-                        "key": "vertical_align",
-                        "default": vertical_align_default,
-                        "label": "Vertical Alignment",
-                        "options": ["top", "center", "bottom"],
-                    },
-                ]
-
-                if isinstance(template_variables, dict):
-                    _insert_overlay_mapping_fields(
-                        template_variables,
-                        align_fields_mapping,
-                        before_keys={"horizontal_offset", "vertical_offset", "builder_level"},
-                    )
-                else:
-                    _insert_overlay_list_fields(
-                        template_variables,
-                        align_fields_list,
-                        before_keys={"horizontal_offset", "vertical_offset", "builder_level"},
-                    )
-
-    return enriched
-
-
-def load_quickstart_overlay_config():
-    return enrich_quickstart_overlay_config(load_quickstart_config("quickstart_overlays.json"))
 
 
 def get_top_imdb_items(library_id, media_type, placeholder_id=None):
@@ -2866,205 +2586,6 @@ def is_imagemaid_running():
         except Exception:
             pass
         return False
-
-
-def get_legacy_custom_fonts_dir() -> Path:
-    return Path(CONFIG_DIR) / "fonts"
-
-
-def get_custom_fonts_dir(config_name: str | None = None) -> Path:
-    if config_name:
-        return get_managed_config_artifact_root(config_name) / "fonts"
-    return get_legacy_custom_fonts_dir()
-
-
-def get_kometa_fonts_dir(kometa_root: Path | None = None) -> Path:
-    if kometa_root is not None:
-        return Path(kometa_root) / "config" / "fonts"
-    return get_kometa_config_dir() / "fonts"
-
-
-def get_font_dirs(include_static: bool = True, include_custom: bool = True, config_name: str | None = None) -> list[Path]:
-    dirs: list[Path] = []
-    seen: set[str] = set()
-
-    if include_custom:
-        custom_paths = []
-        if config_name:
-            custom_paths.append(get_custom_fonts_dir(config_name))
-        custom_paths.extend((get_legacy_custom_fonts_dir(), get_kometa_fonts_dir()))
-        for path in custom_paths:
-            key = str(path)
-            if key not in seen:
-                dirs.append(path)
-                seen.add(key)
-
-    if include_static:
-        for base in (MEIPASS_DIR, BASE_DIR, WORKING_DIR):
-            path = Path(base) / "static" / "fonts"
-            key = str(path)
-            if key not in seen:
-                dirs.append(path)
-                seen.add(key)
-
-    return dirs
-
-
-def list_custom_fonts(config_name: str | None = None) -> list[str]:
-    fonts: set[str] = set()
-    folders = [get_kometa_fonts_dir(), get_legacy_custom_fonts_dir()]
-    if config_name:
-        folders.insert(0, get_custom_fonts_dir(config_name))
-    for folder in folders:
-        if not folder.is_dir():
-            continue
-        for entry in folder.iterdir():
-            if entry.is_file() and entry.suffix.lower() in FONT_EXTENSIONS:
-                fonts.add(entry.name)
-    return sorted(fonts)
-
-
-def list_available_fonts(include_static: bool = True, include_custom: bool = True, config_name: str | None = None) -> list[str]:
-    fonts: set[str] = set()
-    for folder in get_font_dirs(include_static=include_static, include_custom=include_custom, config_name=config_name):
-        if not folder.is_dir():
-            continue
-        for entry in folder.iterdir():
-            if entry.is_file() and entry.suffix.lower() in FONT_EXTENSIONS:
-                fonts.add(entry.name)
-    return sorted(fonts)
-
-
-def migrate_legacy_custom_fonts_to_config(config_name: str | None, font_names: list[str] | tuple[str, ...] | set[str] | None = None) -> dict:
-    try:
-        normalized = require_config_name_for_storage(config_name, context="Config-scoped font migration")
-    except ValueError as exc:
-        return {"copied": [], "skipped": [], "errors": [str(exc)]}
-
-    source_dir = get_legacy_custom_fonts_dir()
-    destination_dir = get_custom_fonts_dir(normalized)
-    copied: list[str] = []
-    skipped: list[str] = []
-    errors: list[str] = []
-
-    if not source_dir.is_dir():
-        return {"copied": copied, "skipped": skipped, "errors": errors}
-
-    requested: set[str] | None = None
-    if font_names is not None:
-        requested = {str(name or "").strip() for name in font_names if str(name or "").strip()}
-        if not requested:
-            return {"copied": copied, "skipped": skipped, "errors": errors}
-
-    destination_dir.mkdir(parents=True, exist_ok=True)
-
-    for entry in sorted(source_dir.iterdir(), key=lambda p: p.name.lower()):
-        if not entry.is_file() or entry.suffix.lower() not in FONT_EXTENSIONS:
-            continue
-        if requested is not None and entry.name not in requested:
-            continue
-        target = destination_dir / entry.name
-        if target.exists():
-            skipped.append(entry.name)
-            continue
-        try:
-            shutil.copy2(entry, target)
-            copied.append(entry.name)
-        except Exception as exc:
-            errors.append(f"Failed to migrate legacy font {entry} -> {target}: {exc}")
-
-    return {"copied": copied, "skipped": skipped, "errors": errors}
-
-
-def sync_custom_fonts(kometa_root: Path | None = None, config_name: str | None = None) -> list[str]:
-    source_dir = get_custom_fonts_dir(config_name)
-    if not source_dir.is_dir():
-        return []
-    dest_dir = get_kometa_fonts_dir(kometa_root)
-    dest_dir.mkdir(parents=True, exist_ok=True)
-    copied: list[str] = []
-    for entry in source_dir.iterdir():
-        if entry.is_file() and entry.suffix.lower() in FONT_EXTENSIONS:
-            shutil.copy2(entry, dest_dir / entry.name)
-            copied.append(entry.name)
-    return copied
-
-
-def collect_font_references(config_data) -> list[str]:
-    fonts: set[str] = set()
-
-    def normalize(value):
-        if isinstance(value, dict):
-            raw = value.get("value")
-            if isinstance(raw, str):
-                return raw.strip()
-            return None
-        if isinstance(value, str):
-            return value.strip()
-        return None
-
-    def walk(obj):
-        if isinstance(obj, dict):
-            for key, val in obj.items():
-                if isinstance(key, str) and (key == "font" or key.endswith("_font")):
-                    norm = normalize(val)
-                    if norm and norm.lower() != "none":
-                        fonts.add(norm)
-                walk(val)
-        elif isinstance(obj, list):
-            for item in obj:
-                walk(item)
-
-    walk(config_data)
-    return sorted(fonts)
-
-
-def copy_fonts_to_kometa(font_refs, kometa_root: Path | None = None, kometa_config_dir: Path | None = None, config_name: str | None = None) -> dict:
-    dest_dir = Path(kometa_config_dir) / "fonts" if kometa_config_dir is not None else get_kometa_fonts_dir(kometa_root)
-    dest_dir.mkdir(parents=True, exist_ok=True)
-    migration = migrate_legacy_custom_fonts_to_config(config_name, font_refs) if config_name else {"copied": [], "skipped": [], "errors": []}
-    sources = get_font_dirs(include_static=True, include_custom=True, config_name=config_name)
-    copied: list[str] = []
-    missing: list[str] = []
-    errors: list[str] = list(migration.get("errors", []))
-
-    for ref in font_refs or []:
-        ref_str = str(ref or "").strip()
-        if not ref_str:
-            continue
-        base = os.path.basename(ref_str)
-        if not base:
-            continue
-
-        source_path = None
-        candidate = Path(ref_str)
-        if candidate.exists():
-            source_path = candidate
-        else:
-            for folder in sources:
-                candidate = Path(folder) / base
-                if candidate.exists():
-                    source_path = candidate
-                    break
-
-        if source_path is None:
-            missing.append(base)
-            continue
-
-        dest_path = dest_dir / base
-        try:
-            if dest_path.resolve() == source_path.resolve():
-                continue
-        except Exception:
-            pass
-
-        try:
-            shutil.copy2(source_path, dest_path)
-            copied.append(base)
-        except Exception as exc:
-            errors.append(f"Failed to copy {source_path} -> {dest_path}: {exc}")
-
-    return {"copied": copied, "missing": missing, "errors": errors}
 
 
 def migrate_config_archives(history_limit: int | None = None) -> dict:
